Shuffled everything around, no functional changes.
